+++ src/Smerge.cpp
@@ -26,7 +26,7 @@ struct Smerge : Module {
 
 	Smerge() {
 		config(NUM_PARAMS, NUM_INPUTS, NUM_OUTPUTS, NUM_LIGHTS);
-		ledDivider.setDivision(4096);
+		ledDivider.setDivision(256);
 		configParam(SORT_PARAM, 0.f, 1.f, 0.f, "Sort voltages");
 	}